revert bigdecimal to float

+++ src/domain/model/src/main/java/com/libraryproviderbackend/text/Text.java
@@ -82,7 +82,7 @@ public class Text extends AggregateRoot<TextId> {
      *
      * @return El precio inicial del texto.
      */
-    public BigDecimal getInitialFloatPrice() {
+    public Float getInitialFloatPrice() {
         return initialPrice.value();
     }
 }
